requirements.txt
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1